Move errors to screen file
