--- VAMobile/src/screens/HomeScreen/VeteranStatusScreen/VeteranStatusError/VeteranStatusError.tsx
@@ -1,74 +1,0 @@
-import React from 'react'
-import { useTranslation } from 'react-i18next'
-
-import { useVeteranStatus } from 'api/veteranStatus'
-import { AlertWithHaptics, ClickToCallPhoneNumber, LargePanel, TextView } from 'components'
-import { NAMESPACE } from 'constants/namespaces'
-import { a11yLabelID, a11yLabelVA } from 'utils/a11yLabel'
-import { displayedTextPhoneNumber } from 'utils/formattingUtils'
-
-const VeteranStatusError = () => {
-  const { data: veteranStatus } = useVeteranStatus()
-  const { t } = useTranslation(NAMESPACE.COMMON)
-  const notConfirmedReason = veteranStatus?.data?.attributes.notConfirmedReason
-
-  const renderError = () => {
-    if (notConfirmedReason === 'NOT_TITLE_38') {
-      return (
-        <AlertWithHaptics
-          variant="warning"
-          header={t('veteranStatus.error.notTitle38.title')}
-          headerA11yLabel={a11yLabelVA(t('veteranStatus.error.notTitle38.title'))}
-          description={t('veteranStatus.error.notTitle38.body1')}
-          descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.notTitle38.body1'))}>
-          <TextView accessibilityLabel={t('veteranStatus.error.notTitle38.body2')}>
-            {t('veteranStatus.error.notTitle38.body2')}
-          </TextView>
-          <ClickToCallPhoneNumber
-            a11yLabel={a11yLabelID(t('8005389552'))}
-            displayedText={displayedTextPhoneNumber(t('8005389552'))}
-            phone={t('8005389552')}
-          />
-        </AlertWithHaptics>
-      )
-    }
-    if (notConfirmedReason === 'ERROR') {
-      return (
-        <AlertWithHaptics
-          variant="error"
-          header={t('errors.somethingWentWrong')}
-          headerA11yLabel={a11yLabelVA(t('errors.somethingWentWrong'))}
-          description={t('veteranStatus.error.generic')}
-          descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.generic'))}
-        />
-      )
-    }
-    return (
-      <AlertWithHaptics
-        variant="warning"
-        header={t('veteranStatus.error.catchAll.title')}
-        headerA11yLabel={a11yLabelVA(t('veteranStatus.error.catchAll.title'))}
-        description={t('veteranStatus.error.catchAll.body')}
-        descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.catchAll.body'))}>
-        <ClickToCallPhoneNumber
-          a11yLabel={a11yLabelID(t('8005389552'))}
-          displayedText={displayedTextPhoneNumber(t('8005389552'))}
-          phone={t('8005389552')}
-        />
-      </AlertWithHaptics>
-    )
-  }
-
-  return (
-    <LargePanel
-      title={t('veteranStatus.title')}
-      rightButtonText={t('close')}
-      dividerMarginBypass={true}
-      testID="veteranStatusTestID"
-      rightButtonTestID="veteranStatusCloseID">
-      {renderError()}
-    </LargePanel>
-  )
-}
-
-export default VeteranStatusError
