+++ VAMobile/src/screens/HomeScreen/VeteranStatusScreen/VeteranStatusScreen.tsx
@@ -13,10 +13,12 @@ import { usePersonalInformation } from 'api/personalInformation/getPersonalInfor
 import { BranchOfService, ServiceData, ServiceHistoryData } from 'api/types'
 import { useVeteranStatus } from 'api/veteranStatus'
 import {
+  AlertWithHaptics,
   BackgroundVariant,
   BorderColorVariant,
   Box,
   BoxProps,
+  ClickToCallPhoneNumber,
   ClickToCallPhoneNumberDeprecated,
   LargePanel,
   MilitaryBranchEmblem,
@@ -25,11 +27,11 @@ import {
 } from 'components'
 import { NAMESPACE } from 'constants/namespaces'
 import { HomeStackParamList } from 'screens/HomeScreen/HomeStackScreens'
+import { a11yLabelID, a11yLabelVA } from 'utils/a11yLabel'
 import { useBeforeNavBackListener, useOrientation, useTheme } from 'utils/hooks'
 import { useReviewEvent } from 'utils/inAppReviews'
 
 import { displayedTextPhoneNumber } from '../../../utils/formattingUtils'
-import VeteranStatusError from './VeteranStatusError/VeteranStatusError'
 
 // import PhotoUpload from 'components/PhotoUpload'
 
@@ -55,18 +57,19 @@ function VeteranStatusScreen({ navigation }: VeteranStatusScreenProps) {
   const combinedPercentText = ratingIsDefined
     ? t('disabilityRating.combinePercent', { combinedPercent: ratingPercent })
     : undefined
+  const branch = mostRecentBranch || ('' as BranchOfService)
 
   useBeforeNavBackListener(navigation, () => {
     registerReviewEvent()
   })
 
   const getPeriodOfService: React.ReactNode = map(serviceHistory, (service: ServiceData) => {
-    const branch = t('militaryInformation.branch', { branch: service.branchOfService })
+    const branchOfService = t('militaryInformation.branch', { branch: service.branchOfService })
     return (
       <Box>
         <Box display="flex" flexDirection="row" alignItems="center" mt={theme.dimensions.condensedMarginBetween}>
           <TextView variant="MobileBody" color="primaryContrast">
-            {branch}
+            {branchOfService}
           </TextView>
         </Box>
         <Box>
@@ -82,7 +85,54 @@ function VeteranStatusScreen({ navigation }: VeteranStatusScreenProps) {
     )
   })
 
-  const branch = mostRecentBranch || ('' as BranchOfService)
+  const getError = () => {
+    const notConfirmedReason = veteranStatus!.data.attributes.notConfirmedReason
+
+    if (notConfirmedReason === 'NOT_TITLE_38') {
+      return (
+        <AlertWithHaptics
+          variant="warning"
+          header={t('veteranStatus.error.notTitle38.title')}
+          headerA11yLabel={a11yLabelVA(t('veteranStatus.error.notTitle38.title'))}
+          description={t('veteranStatus.error.notTitle38.body1')}
+          descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.notTitle38.body1'))}>
+          <TextView accessibilityLabel={t('veteranStatus.error.notTitle38.body2')}>
+            {t('veteranStatus.error.notTitle38.body2')}
+          </TextView>
+          <ClickToCallPhoneNumber
+            a11yLabel={a11yLabelID(t('8005389552'))}
+            displayedText={displayedTextPhoneNumber(t('8005389552'))}
+            phone={t('8005389552')}
+          />
+        </AlertWithHaptics>
+      )
+    }
+    if (notConfirmedReason === 'ERROR') {
+      return (
+        <AlertWithHaptics
+          variant="error"
+          header={t('errors.somethingWentWrong')}
+          headerA11yLabel={a11yLabelVA(t('errors.somethingWentWrong'))}
+          description={t('veteranStatus.error.generic')}
+          descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.generic'))}
+        />
+      )
+    }
+    return (
+      <AlertWithHaptics
+        variant="warning"
+        header={t('veteranStatus.error.catchAll.title')}
+        headerA11yLabel={a11yLabelVA(t('veteranStatus.error.catchAll.title'))}
+        description={t('veteranStatus.error.catchAll.body')}
+        descriptionA11yLabel={a11yLabelVA(t('veteranStatus.error.catchAll.body'))}>
+        <ClickToCallPhoneNumber
+          a11yLabel={a11yLabelID(t('8005389552'))}
+          displayedText={displayedTextPhoneNumber(t('8005389552'))}
+          phone={t('8005389552')}
+        />
+      </AlertWithHaptics>
+    )
+  }
 
   const boxProps: BoxProps = {
     minHeight: 81,
@@ -95,106 +145,110 @@ function VeteranStatusScreen({ navigation }: VeteranStatusScreenProps) {
     borderStyle: 'solid',
   }
 
-  return showError ? (
-    <VeteranStatusError />
-  ) : (
+  return (
     <LargePanel
       title={t('veteranStatus.title')}
       rightButtonText={t('close')}
       dividerMarginBypass={true}
-      removeInsets={true}
+      removeInsets={!showError}
       testID="veteranStatusTestID"
       rightButtonTestID="veteranStatusCloseID">
-      <Box
-        mx={isPortrait ? theme.dimensions.gutter : theme.dimensions.headerHeight}
-        alignItems="center"
-        mt={theme.dimensions.standardMarginBetween}>
-        <VALogo variant="dark" testID="VeteranStatusCardVAIcon" />
-        {/* <Box my={theme.dimensions.standardMarginBetween}>
+      {showError ? (
+        getError()
+      ) : (
+        <>
+          <Box
+            mx={isPortrait ? theme.dimensions.gutter : theme.dimensions.headerHeight}
+            alignItems="center"
+            mt={theme.dimensions.standardMarginBetween}>
+            <VALogo variant="dark" testID="VeteranStatusCardVAIcon" />
+            {/* <Box my={theme.dimensions.standardMarginBetween}>
         //TODO: Put back PhotoUpload later after concerns have been met
           <PhotoUpload width={100} height={100} />
         </Box> */}
-        <Box my={theme.dimensions.formMarginBetween}>
-          <TextView
-            textTransform="capitalize"
-            mb={theme.dimensions.textIconMargin}
-            variant="BitterBoldHeading"
-            color="primaryContrast"
-            testID="veteranStatusFullNameTestID"
-            accessibilityRole="header">
-            {personalInfo?.fullName}
-          </TextView>
-          {accessToMilitaryInfo && (
-            <Box display="flex" flexDirection="row" flexWrap="wrap">
-              <MilitaryBranchEmblem
-                testID="veteranStatusCardBranchEmblem"
-                branch={branch}
-                width={34}
-                height={34}
-                variant="dark"
-              />
-              <TextView ml={10} variant="MobileBody" color="primaryContrast" testID="veteranStatusBranchTestID">
-                {branch}
+            <Box my={theme.dimensions.formMarginBetween}>
+              <TextView
+                textTransform="capitalize"
+                mb={theme.dimensions.textIconMargin}
+                variant="BitterBoldHeading"
+                color="primaryContrast"
+                testID="veteranStatusFullNameTestID"
+                accessibilityRole="header">
+                {personalInfo?.fullName}
               </TextView>
+              {accessToMilitaryInfo && (
+                <Box display="flex" flexDirection="row" flexWrap="wrap">
+                  <MilitaryBranchEmblem
+                    testID="veteranStatusCardBranchEmblem"
+                    branch={branch}
+                    width={34}
+                    height={34}
+                    variant="dark"
+                  />
+                  <TextView ml={10} variant="MobileBody" color="primaryContrast" testID="veteranStatusBranchTestID">
+                    {branch}
+                  </TextView>
+                </Box>
+              )}
             </Box>
-          )}
-        </Box>
-      </Box>
-      <Box mx={isPortrait ? theme.dimensions.gutter : theme.dimensions.headerHeight}>
-        {ratingIsDefined && (
-          <Box {...boxProps}>
-            <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
-              {t('disabilityRating.title')}
-            </TextView>
-            <TextView variant="MobileBody" color="primaryContrast" testID="veteranStatusDisabilityRatingTestID">
-              {combinedPercentText}
-            </TextView>
           </Box>
-        )}
-        <Box {...boxProps} borderBottomWidth={personalInfo?.edipi ? 0 : theme.dimensions.borderWidth}>
-          <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
-            {t('veteranStatus.periodOfService')}
-          </TextView>
-          {getPeriodOfService}
-        </Box>
-        {personalInfo?.edipi && (
-          <Box {...boxProps} borderBottomWidth={theme.dimensions.borderWidth}>
-            <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
-              {t('veteranStatus.dodIdNumber')}
-            </TextView>
-            <TextView variant="MobileBody" color="primaryContrast" testID="veteranStatusDODTestID">
-              {personalInfo?.edipi}
-            </TextView>
+          <Box mx={isPortrait ? theme.dimensions.gutter : theme.dimensions.headerHeight}>
+            {ratingIsDefined && (
+              <Box {...boxProps}>
+                <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
+                  {t('disabilityRating.title')}
+                </TextView>
+                <TextView variant="MobileBody" color="primaryContrast" testID="veteranStatusDisabilityRatingTestID">
+                  {combinedPercentText}
+                </TextView>
+              </Box>
+            )}
+            <Box {...boxProps} borderBottomWidth={personalInfo?.edipi ? 0 : theme.dimensions.borderWidth}>
+              <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
+                {t('veteranStatus.periodOfService')}
+              </TextView>
+              {getPeriodOfService}
+            </Box>
+            {personalInfo?.edipi && (
+              <Box {...boxProps} borderBottomWidth={theme.dimensions.borderWidth}>
+                <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
+                  {t('veteranStatus.dodIdNumber')}
+                </TextView>
+                <TextView variant="MobileBody" color="primaryContrast" testID="veteranStatusDODTestID">
+                  {personalInfo?.edipi}
+                </TextView>
+              </Box>
+            )}
+            <Box my={theme.dimensions.formMarginBetween}>
+              <TextView variant="MobileBody" color="primaryContrast" mb={theme.dimensions.formMarginBetween}>
+                {t('veteranStatus.uniformedServices')}
+              </TextView>
+              <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
+                {t('veteranStatus.fixAnError')}
+              </TextView>
+              <TextView variant="MobileBody" color="primaryContrast" mb={theme.dimensions.condensedMarginBetween}>
+                {t('veteranStatus.fixAnError.2')}
+              </TextView>
+              <ClickToCallPhoneNumberDeprecated
+                phone={t('8008271000')}
+                displayedText={displayedTextPhoneNumber(t('8008271000'))}
+                colorOverride={'veteranStatus'}
+                iconColorOverride={colors.vadsColorWhite}
+              />
+              <TextView variant="MobileBody" color="primaryContrast" my={theme.dimensions.condensedMarginBetween}>
+                {t('veteranStatus.fixAnError.3')}
+              </TextView>
+              <ClickToCallPhoneNumberDeprecated
+                phone={t('8005389552')}
+                displayedText={displayedTextPhoneNumber(t('8005389552'))}
+                colorOverride={'veteranStatus'}
+                iconColorOverride={colors.vadsColorWhite}
+                ttyBypass={true}
+              />
+            </Box>
           </Box>
-        )}
-        <Box my={theme.dimensions.formMarginBetween}>
-          <TextView variant="MobileBody" color="primaryContrast" mb={theme.dimensions.formMarginBetween}>
-            {t('veteranStatus.uniformedServices')}
-          </TextView>
-          <TextView variant="MobileBodyBold" color="primaryContrast" accessibilityRole="header">
-            {t('veteranStatus.fixAnError')}
-          </TextView>
-          <TextView variant="MobileBody" color="primaryContrast" mb={theme.dimensions.condensedMarginBetween}>
-            {t('veteranStatus.fixAnError.2')}
-          </TextView>
-          <ClickToCallPhoneNumberDeprecated
-            phone={t('8008271000')}
-            displayedText={displayedTextPhoneNumber(t('8008271000'))}
-            colorOverride={'veteranStatus'}
-            iconColorOverride={colors.vadsColorWhite}
-          />
-          <TextView variant="MobileBody" color="primaryContrast" my={theme.dimensions.condensedMarginBetween}>
-            {t('veteranStatus.fixAnError.3')}
-          </TextView>
-          <ClickToCallPhoneNumberDeprecated
-            phone={t('8005389552')}
-            displayedText={displayedTextPhoneNumber(t('8005389552'))}
-            colorOverride={'veteranStatus'}
-            iconColorOverride={colors.vadsColorWhite}
-            ttyBypass={true}
-          />
-        </Box>
-      </Box>
+        </>
+      )}
     </LargePanel>
   )
 }
